Add files via upload
Input the Set of nodes. I made it so that it can support only a maximum of 99 nodes for a graph.
The general median problems are classified as NP hard, it would take years for a computer to deal with large numbers of nodes, I keep hearing 50 goes hard on the hardware, no pun intended.